student/requirements.txt
# Student Training System Dependencies

# Core ONNX training framework
onnx>=1.14.0
onnxruntime-training>=1.16.0

# Numerical computing
numpy>=1.24.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0
resampy>=0.4.0

# Database
psycopg2-binary>=2.9.0

# HTTP clients
requests>=2.31.0

# OpenAI API for text generation
openai>=1.0.0

# Configuration
pyyaml>=6.0

# Progress bars and utilities
tqdm>=4.65.0

# Scientific computing
scipy>=1.10.0
